Add StatusError.IsUnknown()
Remove() fails for directories served from a Serve-U server as it
responds with an invalid (Code=24) error code when a remove file
request is sent for a directory.

Attempt a remove directory both for compliant servers that respond with
a failure code, and for non-compliant servers that respond with an
unknown code.